requirements.txt
Django>=3.0,<4.0
psycopg2>=2.7
python-dotenv>=0.19.2
djangorestframework>=3.13.1
Faker>=12.3.0
factory-boy>=3.2.1
pytest>=7.0.1
pytest-django>=4.5.2
pytest-factoryboy>=2.1.0